+++ src/main.py
@@ -7,8 +7,8 @@ import pygame
 from common.move_enum import Move
 from cube.cube import Cube
 from setup import setup_sides
-from solver.side_consts import side_order
 from solver.layer_one import layer_one
+from solver.layer_two.layer_two import layer_two
 from ui.draw_cube import draw_cube
 from ui.square import Square
 
@@ -71,8 +71,11 @@ while True:
                 cube.randomise()
                 current_move = -1
             elif event.key == pygame.K_SPACE:
-                moves = layer_one(cube, True)
+                moves = layer_one.layer_one(cube, True)
                 current_move = 0
+            elif event.key == pygame.K_RALT:
+                layer_two(cube)
+                # current_move = 0
             elif event.key == pygame.K_LALT:
                 cube.sides = deepcopy(setup_sides)
     elapsed = round(time.time() - start_time, 3)
