+++ src/solver/__init__.py
@@ -1,3 +1,6 @@
-__all__ = ["layer_one"]
+__all__ = ["layer_one", "layer_two", "side_data", "side_consts"]
 
+from . import side_consts
+from . import side_data
 from . import layer_one
+from . import layer_two
