+++ src/solver/layer_one/__init__.py
@@ -1,1 +1,8 @@
 __all__ = ["corner_consts", "utils", "corner_consts", "edge_consts", "layer_one", "white_corners", "white_cross"]
+
+from . import layer_one
+from . import corner_consts
+from . import edge_consts
+from . import utils
+from . import white_corners
+from . import white_cross
